Moved store Iter into its own module
